Faster selendroid installer + faster ci.

+++ ci/script.sh
@@ -17,7 +17,7 @@ elif [[ $CI_CONFIG == 'ios' ]]; then
     echo OS X version: `sw_vers -productVersion`
     echo Xcode version: `xcodebuild build -version`
     echo Xcode path: `xcode-select --print-path`
-    ./reset.sh --hardcore --no-npmlink --dev --ios
+    ./reset.sh --no-npmlink --dev --ios
     if [[ $RUN_SAUCE == true ]]; then
         ./ci/upload_build_to_sauce.sh
         TARBALL=sauce-storage:$BZ2_FILE \
@@ -28,7 +28,7 @@ elif [[ $CI_CONFIG == 'ios' ]]; then
 elif [[ $CI_CONFIG == 'android' ]]; then
     source ./ci/android_env
     echo JAVA_HOME: $JAVA_HOME
-    ./reset.sh --hardcore --no-npmlink --dev --android 
+    ./reset.sh --no-npmlink --dev --android 
     if [[ $RUN_SAUCE == true ]]; then
         rm sample-code/apps/ApiDemos
         mv submodules/ApiDemos sample-code/apps/
@@ -38,7 +38,7 @@ elif [[ $CI_CONFIG == 'android' ]]; then
         -p 30 \
         -c android
     fi
-elif [[ $CI_CONFIG == 'others' ]]; then
+elif [[ $CI_CONFIG == 'gappium' ]]; then
     if [[ $TRAVIS_PULL_REQUEST != false ]]; then 
         echo "Skipping this config for pull requests, it takes too long."
         exit 0 
@@ -48,20 +48,36 @@ elif [[ $CI_CONFIG == 'others' ]]; then
     echo Xcode version: `xcodebuild build -version`
     echo Xcode path: `xcode-select --print-path`
     echo JAVA_HOME: $JAVA_HOME
-    ./reset.sh --hardcore --ios --android --no-npmlink
-    ./reset.sh --dev --gappium --selendroid --no-npmlink
+    ./reset.sh --ios --android --selendroid-quick --no-npmlink
+    ./reset.sh --dev --gappium --no-npmlink
     if [[ $RUN_SAUCE == true ]]; then
-        rm sample-code/apps/ApiDemos
-        mv submodules/ApiDemos sample-code/apps/
-        rm sample-code/apps/selendroid-test-app.apk
-        mv submodules/selendroid/selendroid-test-app/target/selendroid-test-app-0.10.0.apk \
-            sample-code/apps/selendroid-test-app.apk
         rm sample-code/apps/io.appium.gappium.sampleapp
         mv submodules/io.appium.gappium.sampleapp sample-code/apps/
         ./ci/upload_build_to_sauce.sh
         TARBALL=sauce-storage:$BZ2_FILE \
         node ./ci/tools/parallel-mocha.js \
-        -p 30 \
-        -c others
+        -p 10 \
+        -c gappium
+    fi
+elif [[ $CI_CONFIG == 'selendroid' ]]; then
+    if [[ $TRAVIS_PULL_REQUEST != false ]]; then 
+        echo "Skipping this config for pull requests, it takes too long."
+        exit 0 
+    fi
+    source ./ci/android_env
+    echo OS X version: `sw_vers -productVersion`
+    echo Xcode version: `xcodebuild build -version`
+    echo Xcode path: `xcode-select --print-path`
+    echo JAVA_HOME: $JAVA_HOME
+    ./reset.sh --android --no-npmlink
+    ./reset.sh --dev --selendroid-quick --no-npmlink
+    if [[ $RUN_SAUCE == true ]]; then
+        rm sample-code/apps/ApiDemos
+        mv submodules/ApiDemos sample-code/apps/
+        ./ci/upload_build_to_sauce.sh
+        TARBALL=sauce-storage:$BZ2_FILE \
+        node ./ci/tools/parallel-mocha.js \
+        -p 10 \
+        -c selendroid
     fi
 fi
